Refactor Layer using the class LayerInteractivity
- Created the class `LayerInteractivity` to manage all the layer
interactions like events and popups.

- `Layer` class moved to a folder called `layer` which contains also:
  - LayerInteractivity class
  - LayerOptions

- Updated Layer imports with the new location of the class.
